Add OCCA_SOURCE_CACHE_DIR environment variable support.
The purpose of this variable is to store source files cache in a separate folder than binary cache.
This feature might be very handful for the projects that don't want to expose sources, but still utilize JIT.
Source cache can be stored in some encrypted tmpfs for example.

Also, this commit exposes occa::env::init so that application can change enviroment variables in a runtime,
and make OCCA reread them.